gui/mdmwebkit: Safely handle potential popen() failure
If we fail to run popen() to obtain the lsb description for any reason,
we should attempt to work around the situation (much like in the previous
commit).

This commit also switches the fixed length within the fgets call to
use sizeof, ensuring we always have the correct size as specified
in the declaration of our lsb_description array.