Changed the way session selection/detection works. Using x-session-manager (default.desktop built-in session) is now a last-resort.
If .dmrc doesn't specify a valid session, we look at DefaultSession in MDM's config.. if it's invalid or set to auto, we proceed to scan for a valid session until we find one. If we don't find any, then and only then we use x-session-manager.

Other changes in this commit include:
  - Stopped saving language selection in mdmlogin and mdmgreeter (in line with mdmwebkit)
  - Removed support for failsafe xterm and gnome sessions
  - Removed a lot of code we no longer need