Append /usr/share/mdm to the XDG_DATA_DIRS variable *after* parsing /etc/profile, otherwise we ended up replacing it with an obsolete version of the variable with the mdm path added.
Instead of in slave.c, just add it to the Xsession default script,
immediately after reading the profile - this ensures we've captured
any other additions as well.

This was preventing flatpak app installs from being recognized by our
menu system (or even working correctly), as a lot of flatpak app info
like .desktop files and icons are kept in /var/lib/flatpak/exports/share.

The flatpak package installs a script to /etc/profile.d to add this
to XDG_DATA_DIRS (which was showing up fine in a tty session, but not
during a user X session).

Note: the Xsession.in file in config/ is not actually used in debian-
based systems, there is a special on in debian that is not generated.
I fixed both, though the debian one has a hardcoded path for now -
I'd rather either ditch the unused one, and generate the 'good' one
in the config folder, or, have both generated (thus avoiding hardcoded
paths) and just pick the correct one in the debian packaging sequence.