Move XDG_DATA_DIR append to *after* run-parts
/etc/X11/Xsession.d/60x11-common_xdg_path sets the default XDG_DATA_DIR
based on some hard-coded paths, and includes an override for setting the DE
path if it's Gnome.

The old placement broke this and /etc/X11/Xsession.d/55cinnamon-session_gnomerc
which tries to do the same thing. Moving the append to after run-parts fixes
this problem, and removes ambiguity.

This problem was introduced in commit 9543455f420d2cada8e021618c535620d08cd6d8.