Don't set XDG_DATA_DIRS in Xsession
Some scripts in /etc/X11/Xsession.d only set that variable if it's empty.
We need to add our path to it later in the sequence and after these scripts are run.
Otherwise, we're missing some default locations in there
and we can end up not being able to launch the session.