mdm: Don't exit on SIGHUP
Make the daemon and the slave ignore SIGHUP calls.

This results in "sudo killall -SIGHUP mdm" having no effect and it should protect the slave from HUP signals.

The SIGHUP signals used by the slave to kill itself (needed for restarts/shutdowns) and communicate with the daemon were replaced with SIGTERM signals.